consolidated all browser-tool related logic under examples
