+++ src/browser.ts
@@ -363,4 +363,3 @@ export default new Ollama()
 
 // export all types from the main entry point so that packages importing types dont need to specify paths
 export * from './interfaces.js'
-export * from './browser-tools.js'
