+++ src/index.ts
@@ -66,6 +66,5 @@ export default new Ollama()
 
 // export all types from the main entry point so that packages importing types dont need to specify paths
 export * from './interfaces.js'
-export * from './browser-tools.js'
 
 export type { AbortableAsyncIterator }
